requirements.txt
clip
clean_fid==0.1.35
einops==0.7.0
einops_exts==0.0.3
filelock==3.12.4
glfw==2.5.5
h5py==3.7.0
hydra-core==1.3.2
imageio==2.23.0
lpips==0.1.4
matplotlib==3.6.2
omegaconf==2.2.3
open3d==0.16.0
opencv_python==4.6.0.66
PyOpenGL==3.1.6
pyrootutils==1.0.4
pytorch_lightning==1.9.4
PyYAML==6.0
scikit_image==0.19.2
setuptools==65.5.0
thop==0.1.1.post2209072238
timm==0.6.12
tqdm==4.64.1
trimesh==3.16.4
ninja==1.11.1.1
torch_geometric==2.4.0
